showcase/docker/stage5_inject/requirements.txt
pandas==1.4.2
numpy==1.22.3
boto3==1.24.0
psutil==5.9.0
requests==2.27.1
python-dotenv==0.20.0
sqlalchemy==1.4.36
pymysql==1.0.2
cryptography==37.0.2
